feat(sms): support for templates for the sms message
