+++ email-templates/en/signin-passwordless-sms/body.txt
@@ -1,0 +1,1 @@
+Your code is ${code}.
