+++ email-templates/fr/signin-passwordless-sms/body.txt
@@ -1,0 +1,1 @@
+Votre code est ${code}.
